+++ backend/selection/src/main/java/com/selection/controller/UserController.java
@@ -6,7 +6,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
@@ -34,19 +33,4 @@ public class UserController {
             .body(articles);
     }
 */
-
-    // Todo: 프론트 연결된 이후엔 삭제할 것
-    @GetMapping("/oauth2/redirect")
-    public ResponseEntity<String> redirectTest(@RequestParam(value = "token")
-        String token) {
-        return ResponseEntity.ok()
-            .body("token : " + token);
-    }
-
-    // Todo: 프론트 연결된 이후엔 삭제할 것
-    @GetMapping("/")
-    public ResponseEntity<String> test() {
-        return ResponseEntity.ok()
-            .body("메인화면 - 누구나 접근 가능");
-    }
 }
